requirements.txt
customtkinter==5.2.0
Pillow==10.0.0
reportlab==4.0.4
sqlite3==2.6.0
Faker==18.10.1
matplotlib==3.7.2